Cover the R2DBC readers and address Sonar findings
Moving the reader logic into AbstractR2dbcResultReader made its typed
read methods and fallback paths count as new code, and most of them had
no direct test. Exercise every read path against both the name and the
ordinal reader with a stub row, including the clob, conversion and bytes
fallbacks and the upper case retries.

Also use a single letter type parameter, an arrow switch, unnamed catch
variables and drop redundant casts.

+++ data-r2dbc/src/main/java/io/micronaut/data/r2dbc/mapper/ColumnOrdinalR2dbcResultReader.java
@@ -40,13 +40,13 @@ final class ColumnOrdinalR2dbcResultReader extends AbstractR2dbcResultReader<Int
     @Override
     @Nullable
     protected Object getValue(Row row, Integer index) {
-        return row.get((int) index);
+        return row.get(index);
     }
 
     @Override
     @Nullable
     protected <T> T getValue(Row row, Integer index, Class<T> type) {
-        return row.get((int) index, type);
+        return row.get(index, type);
     }
 
     @Override
